+++ DelvUI/changelog.md
@@ -1,4 +1,8 @@
 # 1.0.1.1
+Features:
+- Added a sub-option to keep hiding Player Unitframe outside of combat when not full health if "Hide DelvUI outside of combat" is enabled.
+- Added a separate option to always hide Player Unitframe when at full health.
+
 Fixes:
 - Fixed status effects names and descriptions on P4S (English only for now).